trying to center and space the work table
figured out that the tds in the skills table were spaced out because the stars are paragraphs, so inherited the p line height.  Once i set all of the td line height to 2 it worked!  learned to set right margin to 0 if I want to align my sections with left margin.  took out all the cell spacing from html to fully style the tables using css.