Finally centered the profile and ul
Figured out centering and controlling the ul and the profile section was all about margin 0 auto and adjusting the width in PX.  width in % and adjusting padding just made it float around when resizing the browser.

Figured out setting the body to desired mobile view, making it responsive to desktop, and then adjusting individual properties in REM keeps the entire page proportioned   at once and less confusing.

gave the profile a bit of padding for mobile.

The responsiveness now looks amazing!  We'll see how it actually renders on the phone and on the am i responsive webapp