added and styled a bootstrap navbar!
Wow, I had a lot of problems.  My divider borders (hr) suddenly changed.  i had a lot of difficulty positioning and styling the nav and the nav links.

Solutions: I learned about how bootstrap changes stuff around, kind of like the browser defaults.  What I had to do to get my divider borders back was to use the inspect tool to find out what bootstrap had changed, and then override those exact changes in my own stylesheet for the hr.  did a similar thing to style the nav-links.  I discovered I had to write the css for the link changes exactly as written in the inspect tool.  at first I started out typing .nav-links.  It didn't want that.  It wanted .navbar-light to come first.

Was able to position by putting the nav in a div and setting position to absolute and manipulating the top and right borders until it moved to where I wanted it to go.  I also had to play around with the nav-container it to lay on top of the top-container, rather than above it.  Ultimately, I figured out that it had to go inside the top-container.  WHEW!